feat: move models to separate module
