--- bot/src/main/java/edu/java/bot/model/client/AddLinkRequest.java
@@ -1,8 +1,0 @@
-package edu.java.bot.model.client;
-
-import java.net.URI;
-
-public record AddLinkRequest(
-    URI link
-) {
-}
