--- bot/src/main/java/edu/java/bot/model/client/LinkResponse.java
@@ -1,9 +1,0 @@
-package edu.java.bot.model.client;
-
-import java.util.List;
-
-public record LinkResponse(
-    List<LinkResponse> links,
-    int size
-) {
-}
