+++ bot/src/main/java/edu/java/bot/service/LinkUpdater.java
@@ -1,6 +1,6 @@
 package edu.java.bot.service;
 
-import edu.java.bot.model.controller.LinkUpdateRequest;
+import edu.java.LinkUpdateRequest;
 
 public interface LinkUpdater {
     int update(LinkUpdateRequest linkUpdateRequest);
