+++ common/src/main/java/edu/java/ApiErrorResponse.java
@@ -1,4 +1,4 @@
-package edu.java.bot.model.controller.exceptions;
+package edu.java;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
