+++ common/src/main/java/edu/java/LinkResponse.java
@@ -1,4 +1,4 @@
-package edu.java.model.controller;
+package edu.java;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.net.URI;
