+++ common/src/main/java/edu/java/RemoveLinkRequest.java
@@ -1,4 +1,4 @@
-package edu.java.bot.model.client;
+package edu.java;
 
 import java.net.URI;
