+++ scrapper/src/main/java/edu/java/clients/bot/BotClient.java
@@ -1,6 +1,6 @@
 package edu.java.clients.bot;
 
-import edu.java.model.client.LinkUpdateRequest;
+import edu.java.LinkUpdateRequest;
 
 public interface BotClient {
     void sendUpdate(LinkUpdateRequest linkUpdateRequest);
