--- scrapper/src/main/java/edu/java/model/client/LinkUpdateRequest.java
@@ -1,12 +1,0 @@
-package edu.java.model.client;
-
-import java.net.URI;
-import java.util.List;
-
-public record LinkUpdateRequest(
-    Long id,
-    URI url,
-    String description,
-    List<Long> tgChatIds
-) {
-}
