--- scrapper/src/main/java/edu/java/model/controller/RemoveLinkRequest.java
@@ -1,13 +1,0 @@
-package edu.java.model.controller;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonValue;
-import io.swagger.v3.oas.annotations.media.Schema;
-import java.net.URI;
-
-public record RemoveLinkRequest(
-    @Schema(name = "link", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    @JsonProperty("link")
-    URI request
-) {
-}
